Backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
deepface>=0.0.92
bcrypt>=4.1.3
pycryptodome>=3.20.0
opencv-python>=4.9.0.80
scipy>=1.13.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.0
